requirements.txt
fastapi
uvicorn
python-dotenv
aiohttp
mistralai
google-generativeai
anthropic
openai
psutil
netifaces
scapy